# contracts/openapi/solidpos-api-v1.openapi.yaml
openapi: 3.1.0
info:
  title: SolidPOS PosServer API
  version: 1.0.0
  description: 'Runtime contract for SolidPOS PosServer. Macro Phase 30 hardens parity between OpenAPI, implemented Minimal
    API endpoints, request/response contracts and RFC 7807 ProblemDetails errors. Roadmap-only endpoints must not be documented
    here.

    '
servers:
- url: http://localhost:5000/api/v1
  description: Local development
- url: https://api.example.com/api/v1
  description: Production
security:
- bearerAuth: []
...
paths:
  /provisioning/status:
    get:
      tags:
      - Production Provisioning
      summary: Returns production provisioning bootstrap status.
      operationId: get_provisioning_status
      security: []
      responses:
        '200':
          description: Provisioning status.
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/UnexpectedError'
  /provisioning/tenants/bootstrap:
    post:
      tags:
      - Production Provisioning
      summary: Bootstraps a production tenant with admin user and initial store.
      operationId: post_provisioning_tenants_bootstrap
      security: []
      parameters:
      - name: X-SolidPOS-Provision-Key
        in: header
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Production tenant created.
          content:
            application/json:
              schema:
                type: object
        '200':
          description: Idempotent bootstrap already completed.
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/UnexpectedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Credenciales invalidas o token invalido.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: No autorizado por RBAC/policy/tenant context.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Recurso no encontrado.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    ValidationError:
      description: Error de validacion.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Conflict:
      description: Conflicto por version, idempotencia, estado o regla de negocio.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    UnexpectedError:
      description: Error inesperado.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
...
    ProblemDetails:
      type: object
      required:
      - type
      - title
      - status
      - traceId
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        traceId:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...